add Wall Werror to build

+++ dbmse/engine/demo/cross_product/pcrossproductnode.cpp
@@ -74,7 +74,7 @@ bool PCrossProductNode::UpdateRightBlock() {
 }
 
 void PCrossProductNode::Print(size_t indent) {
-  for (int i = 0; i < indent; i++) {
+  for (size_t i = 0; i < indent; i++) {
     cout << " ";
   }
   cout << "CROSS-PRODUCT" << endl;
